model product- fixed

+++ djangoProject/myapp/models.py
@@ -2,7 +2,7 @@ from django.db import models
 
 
 class Product(models.Model):
-    id = models.AutoField(primary_key=True),
-    name = models.CharField(max_length=100),
-    price = models.DecimalField(max_digits=10, decimal_places=2),
+    id = models.AutoField(primary_key=True)
+    name = models.CharField(max_length=100)
+    price = models.FloatField()
     available = models.BooleanField(default=False)
